backend/requirements.txt
Django>=4.0,<5.0
djangorestframework
psycopg2-binary
python-decouple==3.8
django-cors-headers
djangorestframework-simplejwt
dj-database-url
whitenoise
django-environ